pyproject.toml
[project]
name = "drf-practice"
version = "0.1.0"
description = ""
authors = [
    { name = "exkuretrol", email = "33695301+exkuretrol@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "redis[hiredis] (>=5.2.1,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
django-extensions = "^4.1"
ipython = "^9.1.0"
black = "^25.1.0"
pre-commit = "^4.2.0"
werkzeug = "^3.1.3"
django-silk = "^5.3.2"
isort = "^6.0.1"
drf-spectacular = "^0.28.0"

[tool.black]
target-version = ["py310", "py311", "py312"]
preview = true

[tool.isort]
profile = "django"
multi_line_output = 3

[tool.djlint]
profile = "django"
extension = "html,js"
files = "^.*templates/.*.html$"
format_css = false
format_js = true
blank_line_after_tag = "load,extends"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
